+++ src/ngx_http_vhost_traffic_status_display.h
@@ -18,7 +18,7 @@
 #define NGX_HTTP_VHOST_TRAFFIC_STATUS_DISPLAY_CONF_NAME_LEN    1024
 
 /* upper bound of one prometheus line without the node name */
-#define NGX_HTTP_VHOST_TRAFFIC_STATUS_DISPLAY_PROMETHEUS_LINE  128
+#define NGX_HTTP_VHOST_TRAFFIC_STATUS_DISPLAY_PROMETHEUS_LINE 192
 
 /* worst case expansion of ngx_http_vhost_traffic_status_escape_json_pool() */
 #define NGX_HTTP_VHOST_TRAFFIC_STATUS_DISPLAY_JSON_ESCAPE      6